Move send in cli
