+++ apps/cli/src/tools/send/commands/index.ts
@@ -1,0 +1,7 @@
+export * from "./create.command";
+export * from "./delete.command";
+export * from "./edit.command";
+export * from "./get.command";
+export * from "./list.command";
+export * from "./receive.command";
+export * from "./remove-password.command";
