+++ apps/cli/src/tools/send/commands/list.command.ts
@@ -1,10 +1,10 @@
 import { EnvironmentService } from "@bitwarden/common/abstractions/environment.service";
 import { SearchService } from "@bitwarden/common/abstractions/search.service";
-import { SendService } from "@bitwarden/common/abstractions/send/send.service.abstraction";
+import { SendService } from "@bitwarden/common/tools/send/services/send.service.abstraction";
 
-import { Response } from "../../models/response";
-import { ListResponse } from "../../models/response/list.response";
-import { SendResponse } from "../../models/response/send.response";
+import { Response } from "../../../models/response";
+import { ListResponse } from "../../../models/response/list.response";
+import { SendResponse } from "../models/send.response";
 
 export class SendListCommand {
   constructor(
