+++ apps/cli/src/tools/send/index.ts
@@ -1,0 +1,1 @@
+export * from "./commands";
