+++ apps/cli/src/tools/send/models/send-text.response.ts
@@ -1,4 +1,4 @@
-import { SendTextView } from "@bitwarden/common/models/view/send-text.view";
+import { SendTextView } from "@bitwarden/common/tools/send/models/view/send-text.view";
 
 export class SendTextResponse {
   static template(text = "Text contained in the send.", hidden = false): SendTextResponse {
